[fix/#49] navigations 수정

+++ app/src/main/java/com/byeboo/app/presentation/quest/component/QuestCompleteTitle.kt
@@ -18,8 +18,8 @@ import com.byeboo.app.core.designsystem.ui.theme.ByeBooTheme
 
 @Composable
 fun QuestCompleteTitle(
-    stepNumber: Int,
-    questNumber: Int,
+    stepNumber: Long,
+    questNumber: Long,
     createdAt: String,
     questQuestion: String
 ) {
